backend/requirements.txt
annotated-types
anyio
asgiref
certifi
charset-normalizer
colorama
distro
Django
django-cors-headers
djangorestframework
faiss-cpu
h11
httpcore
httpx
idna
jiter
jsonpatch
jsonpointer
langchain>=0.1.0
langchain-classic
langchain-core
langchain-community
langchain-text-splitters
langchain-openai
langsmith
numpy
openai
orjson
ormsgpack
packaging
pydantic
pydantic_core
pypdf
python-docx
python-dotenv
PyYAML
regex
reportlab
requests
requests-toolbelt
sniffio
sqlparse
tenacity
tiktoken
tqdm
typing-inspection
typing_extensions
tzdata
urllib3
uuid_utils
xxhash
